fix modifyLine()
Had to mess with the regular expressions a bit in order to get the rest of those to pass.
We should have all our bases covered, although we won't really know 100% for sure without having to go through all of the lines manually, which would sort of defeat the purpose.